app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
boto3
aws-xray-sdk
python-jose[cryptography]
